// src/app/models/encuesta.ts
export class Encuesta{
    _id?: number;
    titulo: string;
    empresa: string;
    ubicacion: string;
    nombre: string;
    edad: string;
    direccion: string;
    nombreArticulo: string;
    precioArticulo: number;
    marcaArticulo: string;
    pregunta1: string;
    respuesta1: number;
    pregunta2: string;
    respuesta2: number;
    pregunta3: string;
    respuesta3: number;
    pregunta4: string;
    respuesta4: number;
    pregunta5: string;
    respuesta5: number;
    pregunta6: string;
    respuesta6: number;
    pregunta7: string;
    respuesta7: number;
    pregunta8: string;
    respuesta8: number;
    pregunta9: string;
    respuesta9: number;
    pregunta10: string;
    respuesta10: number;
}